AI_Shopping_Jarvis/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.4.0
attrs==25.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
blivet==3.11.0
blivet-gui==2.5.0
boto3==1.35.37
botocore==1.35.37
Brlapi==0.8.5
certifi==2023.5.7
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.1
crypt_r==3.13.1
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
distro==1.9.0
dnf==4.21.1
fedora-third-party==0.10
file-magic==0.4.0
Flask==3.1.0
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.9.0
idna==3.7
initial-setup==0.3.101
iso639==0.1.4
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
langtable==0.0.68
libcomps==0.1.21
libdnf==0.73.3
louis==3.28.0
lxml==5.2.1
MarkupSafe==3.0.2
matplotlib==3.10.1
nftables==0.1
numpy==2.2.4
openai==1.70.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
ply==3.11
productmd==1.38
psutil==5.9.8
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycups==2.0.4
pydantic==2.11.2
pydantic_core==2.33.1
pyenchant==3.2.2
pygdbmi==0.11.0.0
PyGObject==3.48.2
pykickstart==3.58
pyparsing==3.2.1
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-linux-procfs==0.7.3
python-meh==0.52
pytz==2025.1
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
regex==2024.4.28
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.20.0
s3transfer==0.10.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sentry-sdk==2.13.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.8.0
soupsieve==2.6
systemd-python==235
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.1
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xkbregistry==0.3
